quick refactoring -> structure folders, add choose settings in eslint
