+++ src/Routes.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Login from "./components/Login";
+import Login from "./pages/Login";
 import { Redirect } from "react-router";
 import Dashboard from "./pages/Dashboard";
 import Test1 from "./pages/Test1";
